ENH: add LinearSolve(A,X) convenience routine
This commit adds a convenience routine to solve linear systems
of the form `Ax=b` for a given square DenseMatrix instance `A`
and right handside vector, `b`. In general non-trivial case of
a matrix that is 3x3 or greater, this routine serves as a thin
wrapper around the functionality of the LUFactors class that
is provided for convenience. In addition, this commit adds
associated unit tests for the new routine.